transfer shortAmbassador serializer to ambassador app
